Modify commands to standardize inputs
For commands that directly control the train, this standardizes the method signature (args, comms), and allows for args to either be an array for the CLI, or a Hash for the session runner. This is a little clunky and I need to figure out a better way, but it works for now.