Skip file if not JSON parsable and don't require .json extension
Sometimes the session runner thread will pick up and try to read a session file before it has been fully written to disk. This just causes the loop to ignore it if it can't parse the file.

Additionally, the bolt-train-api does not lay down files with a .json extension, so this does not limit the files the session runner reads to .json files.