Add logging
This commits adds logging capabilities to the runner. The logger handles writing strings with the correct format (color) to the console, as well as to a file. In this first pass, the console is always at level 'info' unless the user turns on debug mode. The file is always at level 'debug'. This also moves the configuration file into ~/.bolttrain/config rather than ~/.bolttrain.conf.  The log files will also live in ~/.bolttrain.

This also forks websocket-client-simple and includes it in the project. This is in order to add the logging functionality to the websocket client.